FIX KB-36: Sockets now can send notifications between server and client
